new config format, loading from resources, new model and patt data
Use resource on jar or classpath when no arguments passed.
New format is ARa2Book/0.3, non-URL paths are allowed.
Models and patterns in data and misc directory are the data that they
had been displayed in school festival on Doshisha Univ. held at the end
of Oct 2010, for a program in Doshisha Rohm Plaza.

Signed-off-by: Yuya Tanaka <yuya.tanaka.i@gmail.com>